+++ NEWS.md
@@ -1,5 +1,6 @@
 ### Version 2.2.0
 * Added `heatmap` argument to `plot_gene()`, `plot_region()` and `plot_granges()`. This adds a read-heatmap to the plot.
+* Added `cluster_regions()` function to perform k-means clustering on a table of genomic regions based on methylation profile.
 * Changed averaging method for trends in `plot_gene()`, `plot_region()` and `plot_granges()` to median. This can be changed back using the new `avg_method()` argument.
 
 ### Version 2.0.0